newsletter2paper/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==46.0.1
cssselect2==0.8.0
deprecation==2.1.0
fastapi==0.116.2
fonttools==4.60.1
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
postgrest==2.20.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dotenv==1.1.1
realtime==2.20.0
requests==2.32.5
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlmodel==0.0.25
starlette==0.48.0
storage3==2.20.0
StrEnum==0.4.15
supabase==2.20.0
supabase-auth==2.20.0
supabase-functions==2.20.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
gunicorn==23.0.0
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
zopfli==0.2.3.post1
python-jose==3.3.0
